+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 1.2.0
 * connect timeout fixed (thanks to crazy-rodney, sophisticode, SkuggaEdward, MousyBusiness and cthurston)
+* Add timestamp field to ScanResult class #59 (thanks to simon-iversen)
 
 ## 1.1.3
 * Read RSSI from a connected BLE device #1 (thanks to sophisticode)